+++ tests/Gate.IO.Api.Tests/GlobalUsings.cs
@@ -1,0 +1,5 @@
+global using Gate.IO.Api;
+global using Gate.IO.Api.Alpha;
+global using Newtonsoft.Json;
+global using Newtonsoft.Json.Linq;
+global using Xunit;